Implementação Personagens nas Batalhas Parte 4
-implementação dos personagens criados às Batalhas Parte 4 (esquiva/precisão, feedbacks dos ataques)
-implementação das habilidades às batalhas parte 1

+++ Assets/Scripts/Equipamentos/EquipamentoBase.cs
@@ -24,7 +24,7 @@ public class EquipamentoBase : MonoBehaviour
     public float defesa; //valor da defesa
     public float defesaMagica; //valor da defesa mágica
     public float velocidadeDeMovimento; //valor da velocidade de movimento
-    public float precisao; //valor da precisão
+    public int precisao; //valor da precisão
     public float hp; //valor do hp
     public bool buffConsumivel; //verifica se o equipamento é um consumivel
